# Use ninja for builds.