--- a/python/pyproject.toml
+++ b/python/pyproject.toml
@@ -1,6 +1,6 @@
 
 [build-system]
-requires = ["setuptools>=40.8.0", "wheel", "cmake>=3.18"]
+requires = ["setuptools>=40.8.0", "wheel", "cmake>=3.18", "ninja>=11.1.1"]
 
 [tool.autopep8]
 aggressive = 1