--- a/python/setup.py
+++ b/python/setup.py
@@ -127,13 +127,10 @@ def get_thirdparty_packages(triton_cache_path):
 
 def download_and_copy(src_path, version, url_func):
     base_dir = os.path.dirname(__file__)
-    # src_path = "bin/ptxas"
-    # version = "12.1.105"
     arch = platform.machine()
     if arch == "x86_64":
         arch = "64"
     url = url_func(arch, version)
-    # url = f"https://conda.anaconda.org/nvidia/label/cuda-12.1.1/linux-{arch}/cuda-nvcc-{version}-0.tar.bz2"
     dst_prefix = os.path.join(base_dir, "triton")
     dst_suffix = os.path.join("third_party", "cuda", src_path)
     dst_path = os.path.join(dst_prefix, dst_suffix)
@@ -233,6 +230,7 @@ class CMakeBuild(build_ext):
         # python directories
         python_include_dir = sysconfig.get_path("platinclude")
         cmake_args = [
+            "-G", "Ninja", # Ninja is much faster than make
             "-DCMAKE_EXPORT_COMPILE_COMMANDS=ON",
             "-DLLVM_ENABLE_WERROR=ON",
             "-DCMAKE_LIBRARY_OUTPUT_DIRECTORY=" + extdir,