I noticed that I was not getting much parallelism when building Triton
on my local machine.  After switching from make to ninja, I get a whole
lot more parallelism, and the from-scratch build runs much faster.

On my local machine, I ran:

  $ rm -rf python/build; time pip install -e python

 - Before(make):   999.11s user 124.67s system  484% cpu 3:51.90 total
 - After (ninja): 1539.75s user 187.06s system 2052% cpu 1:24.14 total

On my machine (granted, I have a lot of cores), the speedup is
(3 * 60 + 51) / (1 * 60 + 24) = 2.7x.