@jlebar 2023-09-17
This makes Ninja the one and only build backend.  If folks prefer, we can make it an option.  Thought I'd keep it simple.