@ThomasRaoux 2023-09-17
Let's see if it works on the CI machines but it sounds like Ninja should get installed based on the dependencies.