@jlebar 2023-09-17
CI does not seem to be passing, but I also don't know what the error means.  I get `Error: Process completed with exit code 1.` right after running pre-commit, which itself seems to complete without error?