@ThomasRaoux 2023-09-17
> CI does not seem to be passing, but I also don't know what the error means. I get `Error: Process completed with exit code 1.` right after running pre-commit, which itself seems to complete without error?

It is just the pre-commit formatting checks failing :( 
```
autopep8.................................................................Failed
- hook id: autopep8
- duration: 3.88s
- files were modified by this hook
``` 

You can install pre-commit and run the hook to prevent that:
https://pre-commit.com/