@ThomasRaoux 2023-09-18
> Aha, thanks for fixing it up and merging!
> 

No problem, I wanted to see if it worked on the CI. I do see the build being faster there now :)

> > Merge branch 'main' into build
> 
> That was me conflicting with myself, sorry for the trouble. I have a tool that lets me chain PRs so I don't make merge conflicts with myself, https://github.com/jlebar/git-pr-chain. But (a) it requires me to have push privileges to the main repo, and (b) I'm not sure how well it will work with a workflow where other people modify my branches during review. Anyway we can see.

oh that's cool! I hate that github doesn't have a good way to chain PRs...